feat(suite-native): Mobile Trade: Exchange refactoring and visual updates
