+++ suite-native/module-trading/src/components/exchange/ExchangeFormSkeleton.tsx
@@ -4,12 +4,18 @@ import { SkeletonLargeRow } from '../general/Skeleton/SkeletonLargeRow';
 import { SkeletonSmall } from '../general/Skeleton/SkeletonSmall';
 
 export const ExchangeFormSkeleton = () => (
-    <Card>
-        <VStack>
-            <SkeletonSmall widthPercentage={0.2} />
-            <SkeletonLargeRow leftWidthPercentage={0.4} rightWidthPercentage={0.35} />
-            <SkeletonSmall widthPercentage={0.2} />
-            <SkeletonLargeRow leftWidthPercentage={0.4} rightWidthPercentage={0.3} />
-        </VStack>
-    </Card>
+    <>
+        <Card>
+            <VStack>
+                <SkeletonSmall widthPercentage={0.2} />
+                <SkeletonLargeRow leftWidthPercentage={0.4} rightWidthPercentage={0.35} />
+            </VStack>
+        </Card>
+        <Card>
+            <VStack>
+                <SkeletonSmall widthPercentage={0.2} />
+                <SkeletonLargeRow leftWidthPercentage={0.4} rightWidthPercentage={0.3} />
+            </VStack>
+        </Card>
+    </>
 );
