+++ suite-native/module-trading/src/components/exchange/receive/ExchangeReceiveAmountInput.tsx
