+++ suite-native/module-trading/src/components/exchange/receive/__tests__/ExchangeReceiveAmountInput.comp.test.tsx
@@ -8,8 +8,8 @@ import {
 } from '@suite-native/test-utils';
 
 import { usdcAsset } from '../../../../__fixtures__/tradeableAssets';
-import { useExchangeForm } from '../../../hooks/exchange/useExchangeForm';
-import { ExchangeFormType } from '../../../types/exchange';
+import { useExchangeForm } from '../../../../hooks/exchange/useExchangeForm';
+import { ExchangeFormType } from '../../../../types/exchange';
 import {
     ExchangeReceiveAmountInput,
     ExchangeReceiveAmountInputProps,
