+++ suite-native/module-trading/src/components/exchange/send/ExchangeSendAmountInput.tsx
