+++ suite-native/module-trading/src/components/exchange/send/__tests__/ExchangeSendAmountInput.comp.test.tsx
